buffer_jump: be user friendly when calling LualineBuffersJump
LualineBuffersJump is used to switch between buffers numbered by
lualine. Failing to provide a buffer number that does exist generates
an error that requires user intervention, interrupting the user's flow.
Instead of doing this, use echomsg to inform the user that an attempt
to switch to a non-existent buffer has been attempted, leaving the
state of the buffers intact.

Signed-off-by: Plato Kiorpelidis <kioplato@gmail.com>